+++ src/Symfony/Bundle/FrameworkBundle/Tests/DependencyInjection/Fixtures/php/validation_property_metadata_existence_check.php
@@ -1,7 +1,0 @@
-<?php
-
-$container->loadFromExtension('framework', [
-    'validation' => [
-        'property_metadata_existence_check' => true,
-    ],
-]);