+++ src/Symfony/Component/Validator/Tests/DependencyInjection/RemoveMissingDependenciesPassTest.php
@@ -9,33 +9,45 @@
  * file that was distributed with this source code.
  */
 
-namespace Symfony\Bundle\FrameworkBundle\Tests\DependencyInjection\Compiler;
+namespace Symfony\Component\Validator\Tests\DependencyInjection;
 
 use PHPUnit\Framework\TestCase;
-use Symfony\Bundle\FrameworkBundle\DependencyInjection\Compiler\RemoveUnusedValidatorPropertyInfoLoaderPass;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\PropertyInfo\PropertyInfoExtractor;
+use Symfony\Component\Validator\DependencyInjection\RemoveMissingDependenciesPass;
 use Symfony\Component\Validator\Mapping\Loader\PropertyInfoLoader;
 
-class RemoveUnusedValidatorPropertyInfoLoaderPassTest extends TestCase
+class RemoveMissingDependenciesPassTest extends TestCase
 {
-    public function testTheLoaderIsKeptWhenThePropertyInfoExtractorIsAvailable()
+    public function testThePropertyInfoLoaderIsKeptWhenTheExtractorIsAvailable()
     {
         $container = $this->createContainer();
         $container->register('property_info', PropertyInfoExtractor::class);
 
-        new RemoveUnusedValidatorPropertyInfoLoaderPass()->process($container);
+        new RemoveMissingDependenciesPass()->process($container);
 
         $this->assertTrue($container->hasDefinition('validator.property_info_loader'));
     }
 
-    public function testTheLoaderIsRemovedWhenThePropertyInfoExtractorIsMissing()
+    public function testThePropertyInfoLoaderIsRemovedWhenTheExtractorIsMissing()
     {
         $container = $this->createContainer();
 
-        new RemoveUnusedValidatorPropertyInfoLoaderPass()->process($container);
+        new RemoveMissingDependenciesPass()->process($container);
 
         $this->assertFalse($container->hasDefinition('validator.property_info_loader'));
+    }
+
+    public function testTheConfiguredTranslationDomainWinsOverTheDefault()
+    {
+        $container = $this->createContainer();
+        $container->setParameter('validator.translation_domain', 'validators');
+        $container->setParameter('.validator.translation_domain', 'messages');
+
+        new RemoveMissingDependenciesPass()->process($container);
+
+        $this->assertSame('messages', $container->getParameter('validator.translation_domain'));
+        $this->assertFalse($container->hasParameter('.validator.translation_domain'));
     }
 
     private function createContainer(): ContainerBuilder